Updated value converter
The value converter would check for a number of aliases from the earlier alphas, as the aliases changed a bit, and there was one property editor for a single element and another property editor for a list of elements. The two property editors have now been merged into a single property editor which can be configured to be either a single or multi element type. This makes both the property editor and the value converter much simpler.