Updated the "ElementsController" class and related logic
The "GetImage" and "GetImages" methods have now been updated to support trashed images, as well as honoring the focal point of the returned media. The menthods will initially attempt to look up media in the media cache, but fall back to the media service if not found in the cache (eg. if unpublished).